requirements.txt
requests
python-dotenv
google-generativeai
html2image
openai